user-service/requirements.txt
fastapi
uvicorn
pydantic
pydantic[email]
sqlalchemy
asyncio
pyjwt
loguru
asyncpg
alembic
python-dotenv
